Worked on
Changes on Enemy (Mime)

implementation of animation, currently broken, however, a health system was added to the enemy. The enemy now has a maximum health and a current health value.
When the enemy is hit by a bullet, its health decreases. health bar updates to show the current health as a percentage

The Enemy script was set up to handle damage, a DamageBarMime script was created to manage the enemy’s health bar on

The enemy prefab is set up to be spawned in the game world at specific spawn points, and it has been integrated with the enemy movement and health system.